Add debug for board_f readme grab the value from the line "Reserving 193k for U-Boot at: 2023f000"
and set the value to CONFIG_TEXT_BASE to it
and CONFIG_SYS_UBOOT_START = CONFIG_TEXT_BASE + 0x3fd

this is because there isnt enough space for two u-boots
to be in ram at the same time when reallocating.
I dont know if there is a way to disable reallocating.